+++ src/ExtensibilityToolsPackage.cs
@@ -6,6 +6,8 @@ using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Shell.Interop;
 using Microsoft.VisualStudio.TextTemplating.VSHost;
 using MadsKristensen.ExtensibilityTools.ThemeColorsToolWindow;
+using MadsKristensen.ExtensibilityTools.VsixManifest;
+using MadsKristensen.ExtensibilityTools.VsixManifest.Commands;
 
 namespace MadsKristensen.ExtensibilityTools
 {
@@ -15,6 +17,7 @@ namespace MadsKristensen.ExtensibilityTools
     [ProvideOptionPage(typeof(ExtensibilityOptions), Name, "General", 101, 102, true, new[] { "pkgdef", "vsct" })]
     [ProvideCodeGenerator(typeof(VsctCodeGenerator), VsctCodeGenerator.GeneratorName, VsctCodeGenerator.GeneratorDescription, true, ProjectSystem = ProvideCodeGeneratorAttribute.CSharpProjectGuid)]
     [ProvideCodeGenerator(typeof(VsctCodeGenerator), VsctCodeGenerator.GeneratorName, VsctCodeGenerator.GeneratorDescription, true, ProjectSystem = ProvideCodeGeneratorAttribute.VisualBasicProjectGuid)]
+    [ProvideCodeGenerator(typeof(ResxFileGenerator), ResxFileGenerator.Name, ResxFileGenerator.Desription, true, ProjectSystem = ProvideCodeGeneratorAttribute.CSharpProjectGuid)]
     [ProvideAutoLoad(UIContextGuids80.NoSolution)]
     [Guid(PackageGuids.guidExtensibilityToolsPkgString)]
     [ProvideToolWindow(typeof(SwatchesWindow))]
@@ -28,13 +31,19 @@ namespace MadsKristensen.ExtensibilityTools
         {
             Options = (ExtensibilityOptions)GetDialogPage(typeof(ExtensibilityOptions));
 
+            // VSCT
             AddCustomToolCommand.Initialize(this);
+
+            // Misc
             SignBinaryCommand.Initialize(this);
             ShowProjectInformation.Initialize(this);
             ExportImageMoniker.Initialize(this);
             SwatchesWindowCommand.Initialize(this);
             ShowActivityLog.Initialize(this);
             ToggleVsipLogging.Initialize(this);
+
+            // Vsix Manifest
+            AddResxGeneratorCommand.Initialize(this);
 
             base.Initialize();
         }